+++ adls/ChangeLog.md
@@ -2,7 +2,10 @@
 
 2018.10 Version 11.1.0-preview
 
+* Added `uploadLocalFile` global method in `highlevel.node.ts` to support parallel uploading
+* `PathURL.read` returns a response with a RetriableReadableStream, which will retry when original download stream unexpected ends.
 * [Breaking] `continuation` becomes top1 layer parameters in `listFileSystemsSegment`
+* [Breaking] `contentLength` parameter for `PathURL.update` is `number` type now, previous is `string`
 
 2018.10 Version 11.0.0-preview